+++ tests/hazmat/backends/test_multibackend.py
@@ -197,7 +197,7 @@ class DummyX509Backend(object):
     def load_der_x509_certificate(self, data):
         pass
 
-    def load_pem_x509_request(self, data):
+    def load_pem_x509_csr(self, data):
         pass
 
 
@@ -475,7 +475,7 @@ class TestMultiBackend(object):
 
         backend.load_pem_x509_certificate(b"certdata")
         backend.load_der_x509_certificate(b"certdata")
-        backend.load_pem_x509_request(b"reqdata")
+        backend.load_pem_x509_csr(b"reqdata")
 
         backend = MultiBackend([])
         with raises_unsupported_algorithm(_Reasons.UNSUPPORTED_X509):
@@ -483,4 +483,4 @@ class TestMultiBackend(object):
         with raises_unsupported_algorithm(_Reasons.UNSUPPORTED_X509):
             backend.load_der_x509_certificate(b"certdata")
         with raises_unsupported_algorithm(_Reasons.UNSUPPORTED_X509):
-            backend.load_pem_x509_request(b"reqdata")
+            backend.load_pem_x509_csr(b"reqdata")