+++ CHANGELOG.rst
@@ -19,7 +19,7 @@ Changelog
   and
   :class:`~cryptography.hazmat.primitives.asymmetric.ec.EllipticCurvePrivateKeyWithSerialization`.
 * Add support for parsing X.509 requests with
-  :func:`~cryptography.x509.load_pem_x509_request`.
+  :func:`~cryptography.x509.load_pem_x509_csr`.
 
 0.8.1 - 2015-03-20
 ~~~~~~~~~~~~~~~~~~